requirements.txt
flask
pypdf2
transformers
jieba
pandas
gensim
nltk
pillow
pytesseract
pdf2image
opencv-python
pymongo
flask-restful
flask_cors
torch
torchvision
torchaudio
farm-haystack[inference]
Flask 
Flask-Login 
Flask-Bcrypt 
Flask-SQLAlchemy
cryptography
Flask-MySQL 
mysql-connector-python